% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiageR.R
\name{set_voiage_env}
\alias{set_voiage_env}
\title{Set Python environment for voiage}
\usage{
set_voiage_env(env, type = c("virtualenv", "conda"))
}
\arguments{
\item{env}{The Python environment to use (virtualenv or conda environment name).}

\item{type}{The type of environment ("virtualenv" or "conda").}
}
\value{
None (invisible)
}
\description{
This function sets the Python environment to use for the voiage package.
}
\examples{
\dontrun{
# Use a virtual environment
set_voiage_env("myenv", type = "virtualenv")

# Use a conda environment
set_voiage_env("myenv", type = "conda")
}
}
