% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiageR.R
\docType{package}
\name{voiageR}
\alias{voiageR-package}
\alias{voiageR}
\title{voiageR: R interface to the Rust-backed voiage library}
\description{
Provides package-owned native EVPI and ENBS routines plus optional Python
compatibility paths for EVPPI and EVSI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edithatogo/voiage}
  \item Report bugs at \url{https://github.com/edithatogo/voiage/issues}
}

}
\author{
\strong{Maintainer}: Dylan Mordaunt \email{voiage@users.noreply.github.com} (\href{https://orcid.org/0000-0002-9775-0603}{ORCID})

Authors:
\itemize{
  \item Dylan Mordaunt \email{voiage@users.noreply.github.com} (\href{https://orcid.org/0000-0002-9775-0603}{ORCID})
}

}
\keyword{internal}
